pyproject.toml
[tool.poetry]
name = "datadog-cost-datalake"
version = "1.0.0"
description = "Pipeline de observabilidade de custos Datadog → S3 → Databricks Delta Lake"
authors = ["Ivonaldo Micheluti Dias <ivonaldo.dias@hotmail.com>"]
readme = "README.md"
license = "MIT"
keywords = ["datadog", "finops", "databricks", "delta-lake", "aws-s3", "observability"]

[tool.poetry.dependencies]
python      = "^3.11"
requests    = "^2.31"
boto3       = "^1.34"
pandas      = "^2.2"
pyarrow     = "^15.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest      = "^8.0"
pytest-cov  = "^5.0"
mypy        = "^1.9"
black       = "^24.0"
ruff        = "^0.4"
boto3-stubs = {extras = ["s3"], version = "^1.34"}

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --cov=extractor --cov=loader --cov-report=term-missing"

[tool.mypy]
python_version       = "3.11"
strict               = true
warn_return_any      = true
warn_unused_configs  = true

[tool.black]
line-length    = 100
target-version = ["py311"]

[tool.ruff]
line-length    = 100
target-version = "py311"
select         = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
